CS236Project/requirements.txt
numpy
pandas
matplotlib
Pillow
torch
torchvision
